LayoutTests/platform/ios-simulator/editing/deleting/delete-image-followed-by-two-br-expected.txt

We should be able to delete the br and then the image.
| "
    "
| <img>
|   height="140"
|   width="140"
| <br>
| "This is a line of text.
"
